V0 ready
a few bits and bobs are still missing, but most of it is there. I'm making this library for a project at my job, and all i need it to do for now is detect fields (basic FD pin functionality), which has been done for a while. I intend to finish this library later (probably in my spare time), but i can't promise a hard date.